TUNING: Fix not playing if keyboard opened first
There does not seem to be a value for the keyboard to use if the
keyboard is opened before the slider (or any other thing) is interacted
first and given an ~fslider value. Therefore, at the bottom of the
code, a ~numberfBox.valueAction is added to automatically give this
initial value to the Interpreter and Server.