May 3 2017 Update
TUNING:
- Major clean-up - commented areas for archiving have been deleted
(e.g. removed Axes and harmonic follower indicator on Spectra as has
not been used for a long time and likely a new solution would be
required (needs to be transparent, but sliders currently cannot be)).
Other areas have been moved into module parts.
- The need for an “Ensure it updates” is not necessary, the true
problem was that the array values needed to exist before calling the
functions that used those values. It is now fixed properly.

TECHNICAL DETAILS:
- Begin updating the funding, needs review for specifics.

SPECTRA/KEYBOARD:
- Knob now turns red when it reaches its maximum, indicating that
everything that should be playing on the Keyboard GUI is now red,
regardless of intensity (no matter how low, except if not playing).