Removed the Tuning-testing file
Unnecessary testing, did not uncover why the harmonics dropdown menu
was not properly playing notes, but it is not because of the harmonic
dropdown menu itself. Likely a change due to the ~volume_value (as
fixed previously, the not-loading or nil problem as before in the
frequency slider (once again, moving frequency slider fixes the issue,
so this raises that suspicion).