May 10 2017
TUNING:
- Deviation dot motion works for fundamental = A0 in the positive
direction. It may seem like it works sometimes in the negative
direction, but don’t be fooled - that is either coincidence or somehow
the algorithm rules worked bidirectionally (but that cannot work for
all). A flowchart should be made to keep track of all variables and the
input-output directions (esp. to avoid loops).