May 12 2017
TUNING:
- Moved a calculation for ~deviant_move to above the line where it is
needed (this may be also indicative of the one-click-lag issue with
other, continuous slider systems).
- Commented out unnecessary value calls that were used while developing
the deviator.

deviant_scale_harm:
- Because the starting position is regardless of sign, the if switch
now uses >-2 so it’s always correct positive (~deviant_maximum_harm)
since it’s relying on the distance between the previous harmonic jump
and the fundamental.
- Resolved the ~adder_harm issue: The truncated colour array now starts
from the fundamental and goes up to the previous harmonic. The switch
now does not have overlap so far, but is likely to near the higher
harmonics. The process currently is manual adjustments for exceptions
by using the ~special_sum. Perhaps in the future there will be a more
elegant solution that counts keys; however, this is also an algorithm
based on the fundamental’s pitch class and that in itself is highly
specific.
- Saved a nice format for a 4xif() at the bottom of the script,
commented out.