May 17 2016
menu_5:
- Removed; replaced by a better AppClock function within the Tuning
module itself.

TUNING:
- Added AppClock that activates 0.1s after executed. Sets the harmonic
for deviation to 5. This slight delay is needed so that the menu system
is fully set up before trying to set it to 5, with all its ensuing
functions (e.g. the dropdown and the cover).
- From a previous update, it allows for inspector hover text.
- Added ~inspectorExists variable for proper closing - may replace with
two-state button later.

inspector:
- The Inspector window is fully functional. A specific line of code
must only be added to any object for it to update the text, i.e. as
hover text (~inspectorText).
- Added a section for error messaging (~inspectorError).

SPECTRA:
- Allows for inspector hover text.