May 18 2017
tuning_keyboard_GUI_colours:
- Deleted old unused code.
- Made the harmonic-selected dot grey for white keys. The greyness can
be changed using the ~greyIntensity variable.

TUNING:
- ~harmonic_count variable starts at 8 (the visual dropdown already
did, but now the commanded value used by other systems such as the
deviator now also is 8, which is essential on startup due to the know
reverts to — if 5 > 1, since now it’s 5 !> 8 which allows the system to
work).
- Box shows frequency instead of MIDI note number.
- Moved the ~fundamental_freq so it updates when the fundamental slider
is moved.
- Using a delay, changing the harmonics being played auto-plays.
- Removed a ~numberfBox value preset, doesn’t seem to be needed.

keyboard:
- Now a black dot instead of red.

SPECTRA:
- When “Instruments” is selected, it stops playing.

spectra_screen:
- Being modularized. Interim update.