May 19 2017
TUNING:
- The keyboard button has been moved to the right to make space for the
Autoscale button.

spectra_screen:
- A series of code chunks calculates the fundamental frequency, the
number of harmonics played, the maximum frequency, a surrogate measure
(~sixteenth…), and the appropriate value for the ~spectra_x slider
(autoscales so all harmonics played are viewable in the spectra). Uses
an exponential function, formed through trial and error with point sets
and graphing software.