May 22 2017
TUNING:
- Fixed so the harmonic being deviated resets to the “nil” if it its
number is over the number being played without the inspector needing to
be open (used IF function and in the proper order (call Inspector IF
before setting value back to nil - otherwise the Inspector error would
not be called).

Keyboard:
- Keyboard can be extended, with a button.
- The modularized command for TUNING (needs to be implemented into
SPECTRA) works with the extension.
- Work needs to be done on the proper aesthetics and functionality of
the extension (knowing its limitations).

tuning_keyboard_GUI_colours:
- Updated to work with Keyboard extension.
- Later needs to be updated to be purely-independent, and work only by
loops and algorithms (and be optimized).