June 10 2017
OPENME:
- Added new error for troubleshooting in the page in case new users
encounter a “No space left on device” error. Now under a
“Troubleshooting:” section.

keyboard:
- New images added for the mini-view.
- All C notes now labelled with the pitch and MIDI note.
- New mini-view allows the user to see which octaves are currently in
view in the keyboard window. The highlighted area (in red) shifts along
with view when the octave shift buttons are pressed.
- When the keyboard window is closed by pressing the OS-native close
button (e.g. red x for OSX), the ~key_vis button is also reset. This
makes it more user-friendly and intuitive, so the user can choose to
use the GUI-based ~key_vis button or the OS-native window button.

R:
- Beginning new formulas and functions to integrate R and SC.