June 19 2017
TUNING Module:
- Noticed that when the selected harmonic being deviated is changed
(e.g. from the 3rd to the 2nd), the settings (intensity and deviation)
are preserved. In the future, could reset (intensity of 1 and no
deviation) whenever the selected harmonic being deviated is changed.

SPECTRA Module:
- The volume slider now remains consistent between Sine, Square, Saw,
and Instrument settings (previously it would set itself to a predefined
value every time the pay-setting was changed (e.g. Sine would always
set to a volume of 1, which could be quite unexpected and jolting)).

inspector window for TUNING:
- When the error message appears when the number of harmonics being
played is reduced such that the harmonic selected for deviation is out
of bounds (e.g. 5th selected when only 3 are played), the error message
now disappears after 10 seconds.
- The error message has also been updated to (small grammatical/syntax
improvements over previous version):
"The harmonic selected for deviation was above the number of harmonics
being played. Please select a lower harmonic to deviate or increase the
number of harmonics being played."

inspector window for SPECTRA:
- The number boxes are greyed out to not imply editability (the user
cannot hover over them anyways because once the cursor leaves the
sliders the inspector window clears/resets).

Recording button/ability in TUNING and SPECTRA Modules:
- When recording has started, by quitting the Module (through pressing
the logo to return to the Splashscreen) the recording automatically
stops and saves.
- The line of code at the start of the Modules (s.prepareForRecord;)
allows the recording button to start recording without a delay
(otherwise when the user tries to create the recording it might miss
the important first second or two). However, this has an unfortunate
side-effect of creating a new blank file AIFF (~104b) every time a
Module (either TUNING or SPECTRA) is opened. In the near-future there
should be a better way for setting up recording. These files can be
found in ~/Music/SuperCollider Recordings.