September 8 2017 - Recording Folder and Keyboard Update
Because the "recordings" folder was empty when Git pushed, it did not seem to upload to GitHub. Created a new text file to allow it to upload. Did the same as future-proofing for the R_working_files folder as the .RHistory file is not necessary for MAESTRO but currently included by R programming (usually a hidden file).

Added "~keyboardCodeGui.value" line to the Difference Tones and Synthesizer modules to allow it to load the keyboard GUI file (difference_tones_keyboard_GUI_colours.scd) as needed.

MAPLE_Synthesizer.scd only shows notes on the keyboard if the intensity (i.e. Max. Amp.) is greater than 0 (i.e. is being played/used).